post-review fixes: windowMs overflow, loc-graph ms, enum casing, doc truth
Fixes from the three adversarial reviews of the v57 stack:

- windowMs int32 overflow (regression introduced by the integer bucket
  grid): p.Int is unbounded and the grid cast int32(windowMs) wrapped
  negative -- ?windowMs=4294967295 panicked the row builder (500) and
  made columnar serve 200 with a negative count. resolveWindow now
  rejects windowMs > MaxInt32 (400 invalid_param) and the bucket-count
  guards harden == 0 -> <= 0 (region-control's pattern).

- loc-graph dwell weights flip float64 seconds -> int32 ms, closing the
  surface the pure-ms sweep missed (two reviewers flagged that the
  'every time is ms' headline was false for /loc-graph). The analyzer
  accumulation drops its *0.001 and becomes exact integer math over the
  sample deltas (gap clamp now locgraphSampleDtMs=50); edge/transition
  counts stay counts. /loc-graph gains the timeUnit:"ms" echo
  (LocGraphEnvelope) and the loc-graph artifact echoMs. Goldens
  regenerated; verified byte-identical outside locGraph and exact x1000
  inside. Web tooltips/heatmap divide by 1000 at display. /demoinfo is
  now the genuinely sole seconds island.

- /events and /chat types values validate case-insensitively, matching
  every other token filter (the phase-1 validators had introduced a
  fresh casing island).

- Doc truth: mvd-analytics/README still claimed view outputs 'emit
  seconds' (flat v57 contradiction); ~10 handlers.go param comments and
  the view option-struct godocs said seconds over int32-ms fields;
  events/buckets internal comments described deleted conversions; the
  MCP stream-slice size-guard message said seconds; LosTrack.intervals
  was spec-typed nullable but is never null. API.md + RELEASE_NOTES
  tripwire wording sharpened to be honest: the integer-ms parse rejects
  only non-integer forms -- a whole-number seconds value like from=60
  is valid ms and migrates silently to a 1000x-too-small window; callers
  that passed integer seconds must audit.

+++ mvd-analytics/result/locgraph.go
@@ -17,26 +17,26 @@ type LocGraphResult struct {
 // powerup, so consumers can re-weight the graph / heatmap by combat
 // posture without re-deriving from streams.
 type LocNode struct {
-	Name     string             `json:"name"`
-	X        float32            `json:"x"`
-	Y        float32            `json:"y"`
-	Z        float32            `json:"z"`
-	Total    float64            `json:"total"`
-	ByPlayer map[string]float64 `json:"byPlayer"`
-	ByTeam   map[string]float64 `json:"byTeam,omitempty"`
-	Armed    *LocWeights        `json:"armed,omitempty"`
-	Unarmed  *LocWeights        `json:"unarmed,omitempty"`
-	Quad     *LocWeights        `json:"quad,omitempty"`
-	Pent     *LocWeights        `json:"pent,omitempty"`
+	Name     string           `json:"name"`
+	X        float32          `json:"x"`
+	Y        float32          `json:"y"`
+	Z        float32          `json:"z"`
+	Total    int32            `json:"total"`    // time spent, int32 ms
+	ByPlayer map[string]int32 `json:"byPlayer"` // per-player time, int32 ms
+	ByTeam   map[string]int32 `json:"byTeam,omitempty"` // per-team time, int32 ms
+	Armed    *LocWeights      `json:"armed,omitempty"`
+	Unarmed  *LocWeights      `json:"unarmed,omitempty"`
+	Quad     *LocWeights      `json:"quad,omitempty"`
+	Pent     *LocWeights      `json:"pent,omitempty"`
 }
 
-// LocWeights is a time-spent breakdown (seconds) for one conditioned
+// LocWeights is a time-spent breakdown (int32 ms) for one conditioned
 // metric on a LocNode — same shape as the node's own Total / ByPlayer /
 // ByTeam, omitted entirely when no observed sample met the condition.
 type LocWeights struct {
-	Total    float64            `json:"total"`
-	ByPlayer map[string]float64 `json:"byPlayer"`
-	ByTeam   map[string]float64 `json:"byTeam,omitempty"`
+	Total    int32            `json:"total"`    // time spent, int32 ms
+	ByPlayer map[string]int32 `json:"byPlayer"` // per-player time, int32 ms
+	ByTeam   map[string]int32 `json:"byTeam,omitempty"` // per-team time, int32 ms
 }
 
 // LocEdge is a directed transition from one loc to another, with
